foundry-slag-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiosqlite>=0.19.0
aiofiles>=23.2.0
python-multipart>=0.0.7
rembg[cpu]>=2.0.50
Pillow>=10.0.0
onnxruntime>=1.16.0
